requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.4.0
numpy==1.26.3
lxml==5.1.0
feedparser==6.0.11
duckduckgo-search==5.0.0
python-dotenv==1.0.0